backend/requirements.txt
flask
flask_cors
sqlalchemy
psycopg2-binary
python-dotenv
PyJWT
requests
beautifulsoup4
langchain_community
langchain_openai
langchain-chroma
openai
chromadb
tiktoken
python-pptx
pdfplumber
pytest 
pytest-flask
